refactor(ui): Move form components to shared UI package
This commit extracts common form components from the `ff-site` application and moves them into the shared `@filecoin-foundation/ui` package.

This refactoring centralizes core form elements like `FormField`, `FormInput`, `FormCheckbox`, and their controlled variants into a reusable library. This promotes code reuse, improves consistency across applications, and simplifies future maintenance.

The `ff-site` application has been updated to import these components from the new shared location.
