+++ apps/ff-site/src/app/_components/Form/FormListbox.tsx
@@ -1,3 +1,7 @@
+import {
+  FormField,
+  type FormFieldProps,
+} from '@filecoin-foundation/ui/FormField'
 import { type IconProps } from '@filecoin-foundation/ui/Icon'
 import { Listbox } from '@filecoin-foundation/ui/Listbox'
 import { ListboxButton } from '@filecoin-foundation/ui/Listbox/ListboxButton'
@@ -11,7 +15,6 @@ import {
 } from '@filecoin-foundation/ui/Listbox/ListboxOptions'
 import { type SpacingValue } from '@filecoin-foundation/utils/types/tailwindTypes'
 
-import { FormField, type FormFieldProps } from '@/components/Form/FormField'
 
 export type FormListboxProps = {
   value: OptionType
