+++ apps/ff-site/src/app/_components/Form/FormListboxWithGroups.tsx
@@ -1,10 +1,10 @@
+import { FormField } from '@filecoin-foundation/ui/FormField'
 import { Listbox } from '@filecoin-foundation/ui/Listbox'
 import { ListboxButton } from '@filecoin-foundation/ui/Listbox/ListboxButton'
 import { ListboxGroupHeader } from '@filecoin-foundation/ui/Listbox/ListboxGroupHeader'
 import { ListboxOption } from '@filecoin-foundation/ui/Listbox/ListboxOption'
 import { ListboxOptions } from '@filecoin-foundation/ui/Listbox/ListboxOptions'
 
-import { FormField } from '@/components/Form/FormField'
 import type { FormListboxProps } from '@/components/Form/FormListbox'
 
 export type GroupedOption = {
