+++ packages/ui/src/FormCheckbox/index.ts
@@ -1,0 +1,2 @@
+export * from './FormCheckbox'
+export * from './ControlledFormCheckbox'
