+++ packages/ui/src/FormInput/FormInput.tsx
@@ -1,9 +1,10 @@
-import React from 'react'
-
 import { Input, type InputProps as HeadlessInputProps } from '@headlessui/react'
 import { clsx } from 'clsx'
 
-import { FormField, type FormFieldProps } from '@/components/Form/FormField'
+import {
+  FormField,
+  type FormFieldProps,
+} from '@filecoin-foundation/ui/FormField'
 
 type ExcludedHeadlessUIProps = 'invalid' | 'className'
 
@@ -29,10 +30,7 @@ export function FormInput({
       <Input
         {...rest}
         invalid={Boolean(error)}
-        className={clsx(
-          'block w-full rounded-lg border border-brand-300 bg-brand-800 px-3.5 py-3 placeholder:text-brand-300 hover:border-brand-400 hover:placeholder:text-brand-400 focus:brand-outline focus:text-brand-100 focus:placeholder:text-brand-100 data-disabled:cursor-not-allowed',
-          error && 'border-brand-error',
-        )}
+        className={clsx('form-text-input', error && 'form-text-input-error')}
       />
     </FormField>
   )
