+++ packages/ui/src/FormInput/index.ts
@@ -1,0 +1,2 @@
+export * from './FormInput'
+export * from './ControlledFormInput'
